standardize document on "well suited" (no hyphen)
The need for the hyphen in "well suited" is apparently up for debate,
but some googling finds some support for having the hyphens when
directly before a noun. So   "A well-suited person" but not needed when
saying "That person is well suited for the job"

I don't know for sure, but this changes matches what you had a couple
paragraphs above.